remove quotation marks
Hi there, I've been assigned to evaluate your submission for the data science course on coursera.

The quotation marks ("") were not meant to be part of the file.  
Notice that when you remove the quotation marks, github displays the file with formatting.  ("This is a markdown file" appears in large, bold type).